Cleanup shell handling code
Renames display.rs -> shell.rs and factors code for displaying advisories and
other messages back into main.rs
